requirements.txt
# Original requirements:
openai==1.65.5
gradio==5.29.1
beautifulsoup4
modal==0.74.20
feedparser
rich
python-dotenv
requests
pytest
watchfiles>=0.21.0
hf-xet==1.1.1
fsspec==2024.6.1

# All dependencies (including transitive):
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
distro==1.9.0
fastapi==0.115.12
feedparser==6.0.11
ffmpy==0.5.0
filelock==3.18.0
frozenlist==1.6.0
fsspec==2024.6.1
gradio-client==1.10.1
gradio==5.29.1
groovy==0.1.2
grpclib==0.4.7
h11==0.16.0
h2==4.2.0
hf-xet==1.1.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jinja2==3.1.6
jiter==0.9.0
markdown-it-py==3.0.0
markupsafe==2.1.5
mdurl==0.1.2
modal==0.74.20
multidict==6.4.3
numpy==2.2.5
openai==1.65.5
orjson==3.10.18
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
propcache==0.3.1
protobuf==5.29.4
pydantic-core==2.33.2
pydantic==2.11.4
pydub==0.25.1
pygments==2.19.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pyyaml==6.0.2
requests==2.32.3
rich==14.0.0
ruff==0.11.9
safehttpx==0.1.6
semantic-version==2.10.0
sgmllib3k==1.0.0
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
synchronicity==0.9.12
toml==0.10.2
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.3
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
yarl==1.20.0
